Initial commit.
Many changes committed at the same time. Isn't a good commit. Sorry about that. ¯\_(ツ)_/¯

All UI design refactored to look like this Pinterest's post: https://br.pinterest.com/pin/344243965249378306

Was created a custom UIView to Cards and that card put inside a container to managing the use. This decision was a bad decision because the animation with scale decreases the quality dramatically in the labels and images inside card.
For this project the result is OK

The database was changed. Now Hotel belongs to Flight and Flight is a complete schedule.

Upcoming commits going to be to add comments and removes old code and magic numbers.